requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2018.10.15
chardet==3.0.4
contextlib2==0.5.5
diff-match-patch==20181111
Django==1.10.8
django-admin-list-filter-dropdown==1.0.2
django-celery==3.2.2
django-cors-headers==2.4.0
django-daterange-filter==1.3.0
django-dbbackup==3.2.0
django-filter==2.0.0
django-import-export==1.1.0
django-jwt-auth==0.0.2
django-rest-framework==0.1.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
idna==2.7
image==1.5.27
jdcal==1.4
kombu==3.0.37
numpy==1.15.4
odfpy==1.3.6
openpyxl==2.5.10
pandas==0.23.4
Pillow==5.3.0
PyJWT==1.6.4
python-dateutil==2.7.5
python-resize-image==1.1.18
pytz==2018.7
PyYAML==3.13
raven==6.9.0
requests==2.20.1
simplejson==3.16.0
six==1.11.0
tablib==0.12.1
unicodecsv==0.14.1
urllib3==1.24.1
xlrd==1.1.0
xlwt==1.3.0

